refactor(api): Move PUT /update/vault to its dedicated User's endpoint_controllers file
